Conflicts:
	hybrid_automaton_rlab/src/HybridAutomatonManager.cpp

+++ include/hybrid_automaton/SubjointConfigurationSensor.h
@@ -47,7 +47,7 @@ namespace ha {
             _index.resize(index.size());
             for(int i=0; i<index.rows(); i++)
             {
-                _index.at(i) = index(i,0);
+                _index.at(i) = (int)index(i,0);
             }
         }
